pyproject.toml
[project]
name = "BigData_DataProcess_andMarkdownViewer"
version = "0.1.0"
description = "Your project description goes here"
authors = [
    {name = "shushil girish",email = "girish.sh@northeastern.edu"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "boto3 (>=1.37.11,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "litellm (>=1.63.6,<2.0.0)",
    "anthropic (>=0.49.0,<0.50.0)",
    "pymupdf (>=1.25.4,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.frontend.dependencies]
streamlit = "^1.42.2"
requests = "^2.32.3"


[tool.poetry.group.backend.dependencies]
fastapi = "^0.115.8"
uvicorn = "^0.34.0"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
bcrypt = "^4.2.1"
pandas = "^2.2.3"
pdfplumber = "0.11.4"
pymupdf4llm = "^0.0.17"
boto3 = "^1.36.26"
markitdown = "^0.0.1a4"
apify-client = "^1.9.2"
docling = "^2.24.0"
llama-index = "^0.12.19"

